Rename the `Belin.Cli.Commands` namespace to `Belin.Cli.CommandLine`
