+++ src/command_line/node_command.coffee
