+++ src/command_line/nssm_command.coffee
